// src/Services/API/Identity/API.Identity.Admin.Spa/Client/src/app/core/data/options.ts
import { GroupItem } from "devextreme/data/custom_store";

export interface RestDataStoreUrls {
  key: string
  loadUrl?: string;
  insertUrl?: string;
  updateUrl?: string;
  removeUrl?: string;
  parentID?: number;
  usePutForInsert?: boolean;
}

export type LoadedData<TItem = any> = {
  data: Array<TItem> | Array<GroupItem>;
  totalCount?: number;
  summary?: Array<any>;
  groupCount?: number;
};